elevia-admin-portal/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
beautifulsoup4
playwright
anthropic
chromadb
